# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "企业级任务管理系统的 RESTful API 接口文档",
        "title": "任务管理系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "192.168.12.48:8888",
    "basePath": "/api/v1",
    "paths": {
...
        "/tasks": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取任务列表，支持分页和多条件过滤",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务管理"
                ],
                "summary": "获取任务列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "任务编号（模糊搜索）",
                        "name": "task_no",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "任务标题（模糊搜索）",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "任务类型编码",
                        "name": "task_type_code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "状态编码",
                        "name": "status_code",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "创建人ID",
                        "name": "creator_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "执行人ID",
                        "name": "executor_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "部门ID",
                        "name": "department_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "优先级",
                        "name": "priority",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否在待领池",
                        "name": "is_in_pool",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/dto.PaginationResponse"
                        }
                    },
                    "400": {
                        "description": "参数验证失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建一个新的任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务管理"
                ],
                "summary": "创建任务",
                "parameters": [
                    {
                        "description": "任务信息",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数验证失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tasks/current": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据上下文中的任务ID获取任务基本信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务管理"
                ],
                "summary": "获取当前任务信息",
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/models.Task"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "任务不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tasks/my": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "查询当前登录用户相关的任务，包括自己发布的、自己执行的、被邀请陪审的任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务管理"
                ],
                "summary": "获取我的任务列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "任务编号（模糊搜索）",
                        "name": "task_no",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "任务标题（模糊搜索）",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "任务类型编码",
                        "name": "task_type_code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "状态编码",
                        "name": "status_code",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "优先级",
                        "name": "priority",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "all",
                        "description": "筛选角色：all/creator/executor/jury",
                        "name": "my_role",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功，返回数据中包含 my_role 字段标识用户角色",
                        "schema": {
                            "$ref": "#/definitions/dto.PaginationResponse"
                        }
                    },
                    "400": {
                        "description": "参数验证失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取任务基本信息和当前最新版本的方案、执行计划、以及当前进行中的审核会话",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务详情"
                ],
                "summary": "获取任务详情（包含最新版本的方案和计划）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "任务ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/dto.TaskDetailEnhancedResponse"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "任务不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新任务信息，支持部分字段更新",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务管理"
                ],
                "summary": "更新任务",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "任务ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新的任务信息",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数验证失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "更新失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "软删除任务（有子任务时无法删除）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务管理"
                ],
                "summary": "删除任务",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "任务ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "无效的任务ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "删除失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/tasks/{id}/assign": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "为任务分配执行人，自动更新待领池状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务管理"
                ],
                "summary": "分配执行人",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "任务ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "执行人信息",
                        "name": "assignment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AssignExecutorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "分配成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数验证失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "分配失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/tasks/{id}/transit": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "执行任务状态转换，验证状态有效性并记录变更日志",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务管理"
                ],
                "summary": "执行状态转换",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "任务ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "状态转换信息",
                        "name": "transition",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TaskStatusTransitionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "状态转换成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数验证失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "状态转换失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AssignExecutorRequest": {
            "type": "object",
            "required": [
                "executor_id"
            ],
            "properties": {
                "executor_id": {
                    "description": "执行人用户ID（要分配为任务执行人的用户）",
                    "type": "integer"
                }
            }
        },
...
        "dto.PaginationResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "分页数据列表"
                },
                "page": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "page_size": {
                    "description": "每页数量",
                    "type": "integer"
                },
                "total": {
                    "description": "数据总数",
                    "type": "integer"
                },
                "total_pages": {
                    "description": "总页数",
                    "type": "integer"
                }
            }
        },
...
        "dto.TaskRequest": {
            "type": "object",
            "required": [
                "task_no",
                "task_type_code",
                "title"
            ],
            "properties": {
                "department_id": {
                    "description": "所属部门ID（任务所属的部门，可选）",
                    "type": "integer"
                },
                "description": {
                    "description": "任务描述（详细的任务说明和需求）",
                    "type": "string"
                },
                "executor_id": {
                    "description": "执行人用户ID（任务的具体执行者，可选）",
                    "type": "integer"
                },
                "expected_end_date": {
                    "description": "期望完成日期（任务预计何时完成，可选）",
                    "type": "string"
                },
                "expected_start_date": {
                    "description": "期望开始日期（任务预计何时开始，可选）",
                    "type": "string"
                },
                "is_in_pool": {
                    "description": "是否在待领池（true=未指派执行人，需要其他人认领）",
                    "type": "boolean"
                },
                "parent_task_id": {
                    "description": "父任务ID（用于建立任务层级关系，可选）",
                    "type": "integer"
                },
                "priority": {
                    "description": "任务优先级（1=低，2=中，3=高，4=紧急）",
                    "type": "integer"
                },
                "root_task_id": {
                    "description": "根任务ID（用于快速定位顶层任务，可选）",
                    "type": "integer"
                },
                "solution_deadline": {
                    "description": "思路方案截止时间（仅需求类任务适用，执行人需在此时间前提交解决方案）",
                    "type": "string"
                },
                "status_code": {
                    "description": "任务状态编码（关联 task_statuses.code，可选，默认为初始状态）",
                    "type": "string"
                },
                "task_level": {
                    "description": "任务层级（0=顶层任务，1=一级子任务，以此类推）",
                    "type": "integer"
                },
                "task_no": {
                    "description": "任务编号（唯一，如 REQ-2024-001 格式）",
                    "type": "string"
                },
                "task_type_code": {
                    "description": "任务类型编码（如：需求、缺陷、功能等，关联 task_types.code）",
                    "type": "string"
                },
                "title": {
                    "description": "任务标题（简要描述任务内容）",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}